Polls: show which option the player voted
It's only visible to the player that casted the vote. The other players see what they voted instead.
Had to change the way the poll was being synchronized for this to work. Instead of synchronizing the text, now the poll state is being synchronized and a new system makes sure all clients are connected to changes in the PollState.