Polls - Proof of concept (#29)
* testing-flow

* Allow parametrizing the answer options in the question UI

* Make the poll answers configurable on creation

* The amount of poll answers can be changed on poll creation

* Move setUiRenderer to beginning of main function

Previously, the UI could only be seen only after the player interacted with the cube that creates a poll. For other players that didn't create the poll, we'd like them to be able to see the poll UI without having interacted with the create poll element.

* Synchronize poll

Now, once a poll is created, other players can interact with it.

- Added a PollState to store the poll's replicated state.
- Changed onPointerDown to use a PointerEvents component instead so it can be synced.

* Fix types

* Refactor polls into multiple files

* Poll accumulates answers

* Trying to vote in a poll a second time changes the vote instead of adding a new one

* Polls: show which option the player voted

It's only visible to the player that casted the vote. The other players see what they voted instead.
Had to change the way the poll was being synchronized for this to work. Instead of synchronizing the text, now the poll state is being synchronized and a new system makes sure all clients are connected to changes in the PollState.

* Validate that poll has a title and at least 2 options when creating it

* Fix poll's synchronization issue

Previously polls clicks weren't synchronized correctly. When clicking on a poll, a null poll state was fetched. This change makes it find the correct poll entity.

* Ignore entity-names from linter

That file is autogenerated by the creator hub

* Fix linter issues

---------

Co-authored-by: Giorgio <65674128+Gaffomdz@users.noreply.github.com>