Add message bus
Adds an entity with a synced component that lets players broadcast messages to everyone in the scene.

Messages are typed by the schema defined in MessageBusComponent. If new kind of messages are needed, the constant messageContentAlternatives needs to be changed and a new case should be added to the function handleMessage.