Message Bus (#39)
* Add message bus

Adds an entity with a synced component that lets players broadcast messages to everyone in the scene.

Messages are typed by the schema defined in MessageBusComponent. If new kind of messages are needed, the constant messageContentAlternatives needs to be changed and a new case should be added to the function handleMessage.

* Use MessageBus to broadcast opening the poll result UI

Also needed to change how the poll is identified because the poll registry isn't synchronized.

* Refactor how MessageBusComponentType is obtained

Previously it was a pretty hacky way that required to get the type of an internal variable from BaseComponent