In world polls
Previously, when creating a new poll, a new entity with a textshape, meshcollider and meshrenderer spawned in the world in a random location. Now, a monitor spawns if it wasn't there and when interacting with it, the player can answer the latest poll.

Also, a new entity appears: a button that is used to show the results of the poll.