Forbid double voting in anonymous votes (#82)
This PR allows supporting these cases in anonymous polls:
- Player voted in anonymous poll and tries to vote again:
   - If they voted in the same session, the vote is changed.
   - If they voted but in a different session (like, closing and opening explorer), vote can't change.
   - If they didn't vote in the poll, new vote is casted.

Also, now we are storing the user ids to disable double voting. It's not possible to hide the voters since everything is computed in every client. However, we can mitigate this by storing the voter ids and the votes in different lists, so just by inspecting the results one cannot tell who voted for which option.